pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "syncmind"
version = "0.1.0"
description = "SyncMind: Evaluating Agent Out-of-Sync Recovery"
authors = ["Xuehang Guo <xuehangguo@gmail.com>"]
readme = "README.md"
packages = [
    {include = "syncmind"},
    {include = "syncbench"},
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.2"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
pillow = "^11.0.0"
pyyaml = "^6.0.2"
requests = "^2.31.0"
tqdm = "^4.66.5"
python-dateutil = "^2.9.0"
packaging = "^23.0"
aiohappyeyeballs = "^2.4.3"
aiohttp = "^3.10.10"
aiosignal = "^1.3.1"
annotated-types = "^0.7.0"
anyio = "^4.6.2"
attrs = "^24.2.0"
autopep8 = "^2.3.1"
av = "^13.1.0"
click = "^8.1.7"
contourpy = "^1.3.1"
cycler = "^0.12.1"
distro = "^1.9.0"
fastjsonschema = "*"
frozenlist = "^1.5.0"
fsspec = "^2024.10.0"
gitdb = "^4.0.11"
gitpython = "^3.1.43"
h11 = "^0.14.0"
httpcore = "^1.0.6"
httpx = "^0.27.2"
huggingface-hub = "^0.26.1"
jinja2 = "^3.1.4"
jiter = "^0.6.1"
jsonschema = "^4.23.0"
kiwisolver = "^1.4.7"
litellm = "^1.51.0"
networkx = "^3.4.2"
openai = "^1.52.2"
pydantic = "^2.9.2"
pyparsing = "^3.2.0"
python-dotenv = "^1.0.1"
qwen-vl-utils = "^0.0.8"
regex = "^2024.9.11"
rpds-py = "^0.20.0"
sniffio = "^1.3.1"
tiktoken = "^0.8.0"
tokenizers = "^0.20.1"
typing-extensions = "^4.12.2"
yarl = "^1.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q"
